@ljstella 2024-04-16
Merging into the pydantic2 rewrite via #133 , closing this PR.